pyproject.toml
[project]
name = "todo-integration-ml"
version = "0.1.0"
requires-python = ">=3.11,<3.13"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "celery>=5.3",
  "redis>=5.0",
  "python-dotenv>=1.0",
  "pydantic[email]>=2.7",
  "pydantic-settings>=2.6",
  "SQLAlchemy>=2.0",
  "alembic>=1.13",
  "httpx>=0.27",
  "pyjwt>=2.8",
  "passlib[bcrypt]>=1.7",
  "python-multipart>=0.0.7,<1",

  # --- ML / CSV ---
  "numpy>=1.26,<2.0",
  "pandas>=2.2.2,<3",
  "scikit-learn>=1.4.2,<1.6",
  "joblib>=1.3.2",
  "scipy>=1.10,<2",
]

[project.optional-dependencies]
dev = [
  "black>=24.3",
  "isort>=5.13",
  "ruff>=0.5",
  "mypy>=1.10",
  "pytest>=8.2",
  "pytest-asyncio>=0.24",
  "pytest-cov>=5.0",
  "types-redis>=4.6.0.20240906",
  "types-requests>=2.32.0.20240914"
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 100
force_single_line = false
combine_as_imports = true

[tool.ruff]
line-length = 100
target-version = "py311"
fix = true

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "N", "Q", "W", "C4", "T20", "RET", "SIM", "RUF"]
ignore = ["D"]
extend-ignore = ["RUF003"]

[tool.ruff.lint.isort]
known-first-party = ["src"]

[tool.mypy]
python_version = "3.11"
check_untyped_defs = true
disallow_untyped_defs = true
disallow_any_generics = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_return_any = true
warn_unused_configs = true
pretty = true
show_error_codes = true
exclude = [
  "data/.*",
  "logs/.*",
  "docker/.*",
  ".venv/.*",
]

[[tool.mypy.overrides]]
module = ["celery", "celery.*", "celery.result"]
ignore_missing_imports = true

[tool.pytest.ini_options]
addopts = "-q --cov=./ --cov-report=term-missing --no-header"
testpaths = ["tests"]
xfail_strict = true
filterwarnings = [
  "error::DeprecationWarning",
  "ignore::UserWarning",
]
